Problem:
When components are started, the lookupRealm is used to configure
them. Components are stored using the actual realm (classloader).
This means that, when you have realm A with child realm B,
then component X from realm A could get configured with a component
from realm B, for instance a Logger which is found using a
LoggerManager looked up from B.

When stopping such a component, the same LoggerManager from B
should be used to lookup and unregister the logger.
Since the stop phase executes with the lookuprealm
for the component, which is at that point A, the LoggerManager
can't be found.

Solution:

Store the lookupRealm in the componentmanager and use it
to contextualize the end() phase. This may be used
for other phases too when problems arise there.


UPDATE: wrongly updated ComponentDiscoveryPhase to not
scan parent containers - this was not the right place.
Next commit will roll this back.


git-svn-id: http://svn.codehaus.org/plexus/plexus-containers/trunk@5508 30dfa635-7bf8-0310-9608-e41f2d776689