o component lifecycle manager is now set entirely programmatically and lifecycles can be added programmatically using the container configuration. this will lead to a lot of simplications.
the tests all pass, but I know of a few cases where this setup will not work so I'm going to add some tests, but this pretty much means the abolision of the bootstrap plexus.xml and the container can be setup entirely for the base component. need to add properties and load-on-start components.

git-svn-id: http://svn.codehaus.org/plexus/plexus-containers/trunk@6981 30dfa635-7bf8-0310-9608-e41f2d776689